move slack impl to asset package
