+++ mdw-workflow/assets/com/centurylink/mdw/base/.mdw/versions
@@ -26,7 +26,6 @@ process.jpg=2
 receive.gif=1
 script.gif=1
 send.gif=1
-slack.png=1
 start.png=1
 stop.png=1
 subflow.png=1
@@ -37,6 +36,5 @@ taskOpenEmail.html=9
 taskOpenOneClick.html=6
 velocity.jpg=1
 wait.jpg=1
-web-slack.png=1
 writer.jpg=1
 xml.jpg=1
