+++ mdw-workflow/assets/com/centurylink/mdw/slack/.mdw/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "com.centurylink.mdw.slack",
+  "schemaVersion": "6.0",
+  "version": "6.0.09"
+}
