+++ mdw-workflow/assets/com/centurylink/mdw/slack/.mdw/versions
@@ -1,0 +1,6 @@
+SlackActivity.java=1 Renamed from e:\workspaces\dons\mdw-demo\assets\com\centurylink\mdw\slack\SlackAdapter.java
+SlackAdapter.java=1
+SlackService.java=1
+SlackTaskNotifier.java=1
+TaskNotifier.java=1 Renamed from e:workspacesdonsmdw-demoassetscomcenturylinkmdwslackSlackTaskNotifier.java
+slack.png=1
