+++ mdw-workflow/assets/com/centurylink/mdw/slack/slack.png
